+++ esd/api/esd/parse.go
@@ -4,7 +4,7 @@ import (
 	"encoding/xml"
 	"fmt"
 
-	"github.com/deploymenttheory/winmediafoundry/windowsuup/shared/models"
+	"github.com/deploymenttheory/winmediafoundry/esd/shared/models"
 )
 
 // xmlCatalog mirrors the structure of Microsoft's products.xml: